Add live activities namespace to iOS
