+++ com.onesignal.unity.ios/Runtime/iOSLiveActivitiesManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0831e8d503194992a290b893af97a6a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
